Revert "refacotr($core): separate createCSSRule from core to shared-utils"
This reverts commit 8f89eef61b90834d0a037da3df2633468c1f8e64.
