+++ packages/@vuepress/shared-utils/index.js
@@ -2,7 +2,6 @@ exports.logger = require('./lib/logger')
 exports.env = require('./lib/env')
 exports.writeTemp = require('./lib/writeTemp')
 exports.codegen = require('./lib/codegen')
-exports.webpack = require('./lib/webpack')
 exports.chalk = require('chalk')
 exports.fs = require('fs-extra')
 exports.globby = require('globby')
